@@ -67,18 +76,20 @@
     "date-fns": "^2.30.0",
     "express": "^5.1.0",
     "express-rate-limit": "^7.5.1",
-    "firebase": "^11.10.0",
-    "framer-motion": "^10.16.16",
+    "firebase": "^12.1.0",
+    "firebase-admin": "^13.4.0",
+    "framer-motion": "^11.0.0",
     "helmet": "^7.2.0",
     "lucide-react": "^0.294.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "react-dropzone": "^14.2.3",
     "react-hook-form": "^7.48.2",
     "react-hot-toast": "^2.4.1",
     "react-intersection-observer": "^9.16.0",
-    "react-router-dom": "^6.20.1",
+    "react-router-dom": "^7.0.0",
     "recharts": "^3.1.2",
+    "svg2png-cli": "^1.1.1",
     "yup": "^1.4.0",
     "zustand": "^4.4.7"
   },
@@ -112,17 +127,37 @@
     "jsdom": "^23.0.1",
     "lighthouse": "^12.8.1",
     "msw": "^2.10.5",
-    "postcss": "^8.4.32",
+    "node-fetch": "^3.3.2",
+    "postcss": "^8.5.6",
     "prettier": "^3.1.0",
     "size-limit": "^11.0.0",
-    "tailwindcss": "^3.3.6",
+    "tailwindcss": "^3.4.17",
     "terser": "^5.43.1",
-    "typedoc": "^0.25.4",
-    "typescript": "^5.2.2",
-    "vite": "^7.1.3",
+    "tsx": "^4.20.5",
+    "typedoc": "^0.27.0",
+    "typescript": "^5.7.0",
+    "vite": "^6.0.0",
     "vite-bundle-analyzer": "^0.7.0",
     "vite-plugin-eslint": "^1.8.1",
     "vitest": "^3.2.4"
+  },
+  "overrides": {
+    "@headlessui/react": {
+      "react": "$react",
+      "react-dom": "$react-dom"
+    },
+    "@testing-library/react": {
+      "react": "$react",
+      "react-dom": "$react-dom"
+    },
+    "framer-motion": {
+      "react": "$react",
+      "react-dom": "$react-dom"
+    },
+    "lucide-react": {
+      "react": "$react",
+      "react-dom": "$react-dom"
+    }
   },
   "keywords": [
     "ecommerce",
