@@ -2,17 +2,14 @@ import React, { useState } from 'react';
 import { motion, AnimatePresence } from 'framer-motion';
 import {
   HeartIcon,
-  EyeIcon,
   ShoppingCartIcon,
   StarIcon,
-  CheckBadgeIcon,
 } from '@heroicons/react/24/outline';
 
 import { Product, CarProduct } from '@/types';
 import { useAppStore } from '@/stores/appStore';
 import { HeartSolid } from '@/components/ui/CustomIcons';
 
-import ProductDetailModal from './ProductDetailModal';
 import EnhancedProductDetailsModal from './EnhancedProductDetailsModal';
 import toast from 'react-hot-toast';
 
@@ -30,9 +27,7 @@ const ProductCard = ({
   className = '',
 }: ProductCardProps) => {
   const { language, addToFavorites, removeFromFavorites, isFavorite } = useAppStore();
-  const [isModalOpen, setIsModalOpen] = useState(false);
   const [isEnhancedModalOpen, setIsEnhancedModalOpen] = useState(false);
-  const [imageLoaded, setImageLoaded] = useState(false);
   const [currentImageIndex, setCurrentImageIndex] = useState(0);
 
   const isProductFavorite = isFavorite(product.id);
@@ -69,28 +64,14 @@ const ProductCard = ({
 
     if (!product.inStock) {
       toast.error(language === 'ar' ? 'المنتج غير متوفر' : 'Product is out of stock');
-      //       return;
+      return;
     }
 
     onAddToCart?.(product.id, 1);
     toast.success(language === 'ar' ? 'تم إضافة المنتج للسلة' : 'Added to cart');
   };
 
-  const handleImageError = (): void => {
-    setImageLoaded(false);
-  };
-
-  const handleImageLoad = (): void => {
-    setImageLoaded(true);
-  };
-
-  const nextImage = (e: React.MouseEvent<HTMLButtonElement>): void => {
-    e.preventDefault();
-    e.stopPropagation();
-    setCurrentImageIndex((prev: number) => (prev + 1) % product.images.length);
-  };
-
-  const prevImage = (e: React.MouseEvent<HTMLButtonElement>): void => {
+  const handlePreviousImage = (e: React.MouseEvent<HTMLButtonElement>): void => {
     e.preventDefault();
     e.stopPropagation();
     setCurrentImageIndex(
@@ -98,263 +79,202 @@ const ProductCard = ({
     );
   };
 
-  return (
-    <>
-      <motion.div
-        className={`group relative bg-white rounded-xl shadow-lg hover:shadow-xl transition-all duration-300 overflow-hidden border border-neutral-200 hover:border-primary-300 ${className}`}
-        whileHover={{ y: -4, scale: 1.02 }}
-        initial={{ opacity: 0, y: 20 }}
-        animate={{ opacity: 1, y: 0 }}
-        transition={{ duration: 0.3 }}
-      >
-        {/* Image Section */}
-        <div className='relative h-48 bg-neutral-100 overflow-hidden'>
-          <div className='absolute inset-0 bg-gradient-to-t from-black/20 to-transparent opacity-0 group-hover:opacity-100 transition-opacity duration-300' />
+  const handleNextImage = (e: React.MouseEvent<HTMLButtonElement>): void => {
+    e.preventDefault();
+    e.stopPropagation();
+    setCurrentImageIndex((prev: number) => (prev + 1) % product.images.length);
+  };
 
-          {/* Main Image */}
-          <AnimatePresence mode='wait'>
+  return (
+    <div className={`bg-white rounded-xl shadow-sm hover:shadow-lg transition-all duration-300 flex flex-col overflow-hidden border border-neutral-200 group ${className || ''}`}>
+      {/* Image Section - Fixed aspect ratio with responsive scaling */}
+      <div className="relative w-full">
+        <div className="aspect-[4/3] w-full bg-neutral-100 overflow-hidden">
+          <AnimatePresence mode="wait">
             <motion.img
               key={currentImageIndex}
               src={product.images[currentImageIndex]?.url || primaryImage?.url}
-              alt={product.images[currentImageIndex]?.alt || primaryImage?.alt}
-              className='w-full h-full object-cover transition-transform duration-500 group-hover:scale-110'
-              onLoad={handleImageLoad}
-              onError={handleImageError}
-              initial={{ opacity: 0 }}
-              animate={{ opacity: imageLoaded ? 1 : 0 }}
-              exit={{ opacity: 0 }}
-              loading='lazy'
+              alt={product.title}
+              className="w-full h-full object-cover object-center transition-transform duration-300 group-hover:scale-105"
+              initial={{ opacity: 0, scale: 1.05 }}
+              animate={{ opacity: 1, scale: 1 }}
+              exit={{ opacity: 0, scale: 0.95 }}
+              transition={{ duration: 0.2 }}
             />
           </AnimatePresence>
-
-          {/* Image Navigation */}
+          
+          {/* Gradient Overlay */}
+          <div className="absolute inset-0 bg-gradient-to-t from-black/20 to-transparent opacity-0 group-hover:opacity-100 transition-opacity duration-300" />
+          
+          {/* Navigation Buttons - Positioned for desktop scaling */}
           {product.images.length > 1 && (
-            <div className='absolute inset-0 flex items-center justify-between opacity-0 group-hover:opacity-100 transition-opacity duration-300'>
+            <>
               <button
-                onClick={prevImage}
-                className='ml-2 p-1.5 bg-white/80 hover:bg-white rounded-full shadow-md backdrop-blur-sm transition-all duration-200'
-                type='button'
-                aria-label='Previous image'
+                onClick={handlePreviousImage}
+                className="absolute left-2 top-1/2 -translate-y-1/2 bg-white/90 backdrop-blur-sm p-1.5 rounded-full shadow-md opacity-0 group-hover:opacity-100 transition-all duration-200 hover:bg-white hover:scale-110"
+                aria-label="Previous image"
               >
-                <svg
-                  className='w-4 h-4 text-neutral-700'
-                  fill='none'
-                  stroke='currentColor'
-                  viewBox='0 0 24 24'
-                >
-                  <path
-                    strokeLinecap='round'
-                    strokeLinejoin='round'
-                    strokeWidth={2}
-                    d='M15 19l-7-7 7-7'
-                  />
+                <svg className="w-3.5 h-3.5 text-neutral-700" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M15 19l-7-7 7-7" />
                 </svg>
               </button>
               <button
-                onClick={nextImage}
-                className='mr-2 p-1.5 bg-white/80 hover:bg-white rounded-full shadow-md backdrop-blur-sm transition-all duration-200'
-                type='button'
-                aria-label='Next image'
+                onClick={handleNextImage}
+                className="absolute right-2 top-1/2 -translate-y-1/2 bg-white/90 backdrop-blur-sm p-1.5 rounded-full shadow-md opacity-0 group-hover:opacity-100 transition-all duration-200 hover:bg-white hover:scale-110"
+                aria-label="Next image"
               >
-                <svg
-                  className='w-4 h-4 text-neutral-700'
-                  fill='none'
-                  stroke='currentColor'
-                  viewBox='0 0 24 24'
-                >
-                  <path
-                    strokeLinecap='round'
-                    strokeLinejoin='round'
-                    strokeWidth={2}
-                    d='M9 5l7 7-7 7'
-                  />
+                <svg className="w-3.5 h-3.5 text-neutral-700" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 5l7 7-7 7" />
                 </svg>
               </button>
-            </div>
+            </>
           )}
-
-          {/* Image Loading Placeholder */}
-          {!imageLoaded && (
-            <div className='absolute inset-0 flex items-center justify-center bg-neutral-100'>
-              <div className='w-8 h-8 border-2 border-primary-500 border-t-transparent rounded-full animate-spin' />
-            </div>
-          )}
-        </div>
-
-        {/* Content Section */}
-        <div className='p-4 space-y-3'>
-          {/* Title and Rating */}
-          <div className='space-y-2'>
-            <div className='flex items-start justify-between'>
-              <h3 className='font-semibold text-neutral-900 line-clamp-2 text-sm leading-tight group-hover:text-primary-600 transition-colors'>
-                {product.title}
-              </h3>
-            </div>
-
-            <div className='flex items-center justify-between text-xs text-neutral-600'>
-              <div className='flex items-center space-x-1'>
-                <div className='flex items-center'>
-                  {[...Array(5)].map((_, i) => (
-                    <StarIcon
-                      key={i}
-                      className={`w-3 h-3 ${
-                        i < Math.floor(product.rating)
-                          ? 'text-yellow-400 fill-current'
-                          : 'text-neutral-300'
-                      }`}
-                    />
-                  ))}
-                </div>
-                <span className='text-neutral-500'>({product.reviewCount})</span>
-              </div>
-
-              <div className='flex items-center space-x-3 text-neutral-500'>
-                <div className='flex items-center space-x-1'>
-                  <EyeIcon className='w-3 h-3' />
-                  <span>{product.views}</span>
-                </div>
-                <div className='flex items-center space-x-1'>
-                  <HeartIcon className='w-3 h-3' />
-                  <span>{product.favorites}</span>
-                </div>
-              </div>
-            </div>
+          
+          {/* Image Indicators */}
+          <div className="absolute bottom-2 left-1/2 -translate-x-1/2 flex space-x-1">
+            {product.images.map((_, index) => (
+              <button
+                key={index}
+                onClick={() => setCurrentImageIndex(index)}
+                className={`w-1.5 h-1.5 rounded-full transition-all duration-200 ${
+                  index === currentImageIndex
+                    ? 'bg-white scale-125'
+                    : 'bg-white/60 hover:bg-white/80'
+                }`}
+                aria-label={`Go to image ${index + 1}`}
+              />
+            ))}
           </div>
 
-          {/* Car Details */}
-          {carDetails && (
-            <div className='grid grid-cols-2 gap-2 text-xs bg-neutral-50 p-2 rounded-lg'>
-              <div className='flex justify-between'>
-                <span className='text-neutral-600'>{language === 'ar' ? 'السنة' : 'Year'}:</span>
-                <span className='font-medium'>{carDetails.year}</span>
-              </div>
-              <div className='flex justify-between'>
-                <span className='text-neutral-600'>
-                  {language === 'ar' ? 'الكيلومترات' : 'Mileage'}:
-                </span>
-                <span className='font-medium'>{carDetails.mileage?.toLocaleString()} km</span>
-              </div>
-              <div className='col-span-2 flex justify-between'>
-                <span className='text-neutral-600'>{language === 'ar' ? 'الطراز' : 'Model'}:</span>
-                <span className='font-medium'>
-                  {carDetails.make} {carDetails.model}
-                </span>
-              </div>
-            </div>
-          )}
+          {/* Stock Status Badge */}
+          <div className="absolute top-2 left-2">
+            <span
+              className={`inline-flex items-center px-2 py-0.5 rounded-full text-xs font-medium ${
+                product.inStock
+                  ? 'bg-green-100 text-green-800'
+                  : 'bg-red-100 text-red-800'
+              }`}
+            >
+              {product.inStock ? 'In Stock' : 'Out of Stock'}
+            </span>
+          </div>
 
-          {/* Features */}
-          {product.features && product.features.length > 0 && (
-            <div className='flex flex-wrap gap-1'>
-              {product.features.slice(0, 3).map((feature, index) => (
-                <span
-                  key={index}
-                  className='px-2 py-1 bg-neutral-100 text-neutral-600 text-xs rounded-full'
-                >
-                  {feature}
-                </span>
+          {/* Favorite Button */}
+          <button
+            onClick={handleToggleFavorite}
+            className="absolute top-2 right-2 bg-white/90 backdrop-blur-sm p-1.5 rounded-full shadow-md transition-all duration-200 hover:scale-110"
+            aria-label={isProductFavorite ? 'Remove from favorites' : 'Add to favorites'}
+          >
+            {isProductFavorite ? (
+              <HeartSolid className="w-4 h-4 text-red-500" />
+            ) : (
+              <HeartIcon className="w-4 h-4 text-neutral-600 hover:text-red-500" />
+            )}
+          </button>
+        </div>
+      </div>
+
+      {/* Content Section - Optimized for desktop scaling */}
+      <div className="p-3 sm:p-4 flex flex-col flex-grow space-y-2">
+        {/* Title and Rating */}
+        <div className="space-y-1.5">
+          <h3 className="text-base sm:text-lg font-semibold text-neutral-900 line-clamp-2 leading-tight">
+            {product.title}
+          </h3>
+          
+          <div className="flex items-center space-x-1">
+            <div className="flex items-center">
+              {[...Array(5)].map((_, i) => (
+                <StarIcon
+                  key={i}
+                  className={`w-3.5 h-3.5 ${
+                    i < Math.floor(product.rating || 0)
+                      ? 'text-yellow-400 fill-yellow-400'
+                      : 'text-neutral-300'
+                  }`}
+                />
               ))}
-              {product.features.length > 3 && (
-                <span className='px-2 py-1 bg-neutral-100 text-neutral-600 text-xs rounded-full'>
-                  +{product.features.length - 3}
-                </span>
-              )}
             </div>
-          )}
-
-          {/* Price Section */}
-          <div className='flex items-center justify-between'>
-            <div className='space-y-1'>
-              <div className='flex items-center space-x-2'>
-                <span className='text-lg font-bold text-primary-600'>
-                  {formatPrice(product.price)}
-                </span>
-                {product.originalPrice && product.originalPrice > product.price && (
-                  <span className='text-sm text-neutral-500 line-through'>
-                    {formatPrice(product.originalPrice)}
-                  </span>
-                )}
-              </div>
-
-              {/* Stock Status */}
-              <div className='flex items-center text-xs'>
-                {product.inStock ? (
-                  <>
-                    <CheckBadgeIcon className='w-3 h-3 text-green-500 mr-1' />
-                    <span className='text-green-600'>
-                      {language === 'ar' ? 'متوفر' : 'In Stock'}
-                    </span>
-                  </>
-                ) : (
-                  <span className='text-red-600'>
-                    {language === 'ar' ? 'غير متوفر' : 'Out of Stock'}
-                  </span>
-                )}
-              </div>
-            </div>
-
-            {/* Action Buttons */}
-            <div className='flex items-center space-x-2'>
-              {/* Details Button */}
-              <motion.button
-                onClick={() => setIsEnhancedModalOpen(true)}
-                className='px-3 py-1.5 bg-slate-100 hover:bg-slate-200 text-slate-700 text-xs font-medium rounded-lg transition-colors'
-                whileHover={{ scale: 1.05 }}
-                whileTap={{ scale: 0.95 }}
-                type='button'
-              >
-                {language === 'ar' ? 'التفاصيل' : 'Details'}
-              </motion.button>
-
-              <motion.button
-                onClick={handleToggleFavorite}
-                className='p-2 rounded-full hover:bg-neutral-100 transition-colors'
-                whileHover={{ scale: 1.1 }}
-                whileTap={{ scale: 0.9 }}
-                type='button'
-                aria-label={isProductFavorite ? 'Remove from favorites' : 'Add to favorites'}
-              >
-                {isProductFavorite ? (
-                  <HeartSolid className='w-5 h-5 text-red-500' />
-                ) : (
-                  <HeartIcon className='w-5 h-5 text-neutral-400' />
-                )}
-              </motion.button>
-
-              {product.category !== 'services' && product.inStock && (
-                <motion.button
-                  onClick={handleAddToCart}
-                  className='p-2 bg-primary-500 text-white rounded-full hover:bg-primary-600 transition-colors'
-                  whileHover={{ scale: 1.1 }}
-                  whileTap={{ scale: 0.9 }}
-                  type='button'
-                  aria-label='Add to cart'
-                >
-                  <ShoppingCartIcon className='w-5 h-5' />
-                </motion.button>
-              )}
-            </div>
+            <span className="text-xs text-neutral-600">
+              {product.rating?.toFixed(1) || '0.0'} ({product.reviewCount || 0})
+            </span>
           </div>
         </div>
 
-        {/* Verified Badge */}
-        {product.status === 'published' && (
-          <div className='absolute top-0 right-0 bg-green-500 text-white p-1 rounded-bl-lg'>
-            <CheckBadgeIcon className='w-4 h-4' />
+        {/* Car Details */}
+        <div className="flex flex-wrap gap-1.5 text-xs text-neutral-600">
+          {carDetails?.year && (
+            <span className="bg-neutral-100 px-1.5 py-0.5 rounded">{carDetails.year}</span>
+          )}
+          {carDetails?.mileage && (
+            <span className="bg-neutral-100 px-1.5 py-0.5 rounded">
+              {carDetails.mileage.toLocaleString()} km
+            </span>
+          )}
+          {product.fuelType && (
+            <span className="bg-neutral-100 px-1.5 py-0.5 rounded capitalize">{product.fuelType}</span>
+          )}
+        </div>
+
+        {/* Features */}
+        {product.features && product.features.length > 0 && (
+          <div className="flex flex-wrap gap-1">
+            {product.features.slice(0, 3).map((feature, index) => (
+              <span
+                key={index}
+                className="text-xs bg-primary-50 text-primary-700 px-1.5 py-0.5 rounded capitalize"
+              >
+                {feature}
+              </span>
+            ))}
+            {product.features.length > 3 && (
+              <span className="text-xs text-neutral-500">+{product.features.length - 3}</span>
+            )}
           </div>
         )}
-      </motion.div>
 
-      {/* Product Detail Modal */}
-      {isModalOpen && (
-        <ProductDetailModal
-          product={product}
-          isOpen={isModalOpen}
-          onClose={() => setIsModalOpen(false)}
-          onAddToCart={onAddToCart}
-          onToggleFavorite={onToggleFavorite}
-        />
-      )}
+        {/* Price Section */}
+        <div className="mt-auto pt-2">
+          <div className="flex items-end justify-between">
+            <div>
+              <p className="text-lg sm:text-xl font-bold text-neutral-900">
+                {formatPrice(product.price)}
+              </p>
+              {product.originalPrice && product.originalPrice > product.price && (
+                <p className="text-xs text-neutral-500 line-through">
+                  {formatPrice(product.originalPrice)}
+                </p>
+              )}
+            </div>
+            <span className={`text-xs px-1.5 py-0.5 rounded-full ${
+              product.condition === 'new'
+                ? 'bg-green-100 text-green-800'
+                : product.condition === 'used'
+                ? 'bg-orange-100 text-orange-800'
+                : 'bg-blue-100 text-blue-800'
+            }`}>
+              {product.condition}
+            </span>
+          </div>
+        </div>
+
+        {/* Action Buttons */}
+        <div className="flex gap-1.5 pt-2">
+          <button
+            onClick={() => setIsEnhancedModalOpen(true)}
+            className="flex-1 bg-neutral-100 text-neutral-700 px-2 py-1.5 rounded-md text-xs font-medium hover:bg-neutral-200 transition-colors"
+          >
+            Quick View
+          </button>
+          <button
+            onClick={handleAddToCart}
+            disabled={!product.inStock}
+            className="flex-1 bg-primary-600 text-white px-2 py-1.5 rounded-md text-xs font-medium hover:bg-primary-700 transition-colors disabled:bg-neutral-300 disabled:cursor-not-allowed"
+          >
+            {product.inStock ? 'Add to Cart' : 'Out of Stock'}
+          </button>
+        </div>
+      </div>
 
       {/* Enhanced Product Details Modal */}
       {isEnhancedModalOpen && (
@@ -379,7 +299,7 @@ const ProductCard = ({
           isFavorite={isProductFavorite}
         />
       )}
-    </>
+    </div>
   );
 };
 
